唯一值BitMap合并。
